requirements.txt
python-telegram-bot==20.8
aiosqlite
requests
openai
python-dotenv
nest_asyncio
